Restructure combat classes for future additions
